relocate reducers
