+++ src/store.js
@@ -3,7 +3,7 @@ import { createStore
     , applyMiddleware
 } from 'redux'
 import thunk from 'redux-thunk';
-import rootReducer from './reducer'
+import rootReducer from './reducers';
 
 const store = createStore(rootReducer, compose(applyMiddleware(thunk), window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()))
